+++ src/ssr-state-service/provider.ts
@@ -1,0 +1,4 @@
+import { provider } from '../provider'
+import { SsrStateService } from './ssr-state-service'
+
+export const useSsrState = provider (() => new SsrStateService())