+++ src/types.ts
@@ -2,8 +2,17 @@ import { DescriptorsContainer } from './plugin/descriptors-container'
 
 export type DependencyFactory<Target> = (...args: any[]) => Target
 
-export type Provider<Target> = () => Target
+export interface Provider<Target> {
+  (): Target
+  readonly asKey: symbol
+}
+  
 
 export interface DependencyContainerPlugin {
   readonly dependencyContainer: DescriptorsContainer
+}
+
+export interface ProviderOptions {
+  persistentInstance?: boolean
+  key?: symbol
 }